Updates, fixes, etc
this S1F and S2A support is making things really confusing

+++ Project Files/Sonic 2 Absolute/SonLVLObjDefs/Enemies/Grabber.cs
@@ -39,7 +39,7 @@ namespace S2ObjectDefinitions.Enemies
 					{ "Right", 1 }
 				},
 				(obj) => (obj.PropertyValue == 0) ? 0 : 1,
-				(obj, value) => obj.PropertyValue = (byte)((obj.PropertyValue & ~1) | (byte)((int)value)));
+				(obj, value) => obj.PropertyValue = (byte)((obj.PropertyValue & ~1) | (int)value));
 		}
 
 		public override ReadOnlyCollection<byte> Subtypes
