+++ Project Files/Sonic 2 Absolute/SonLVLObjDefs/Enemies/Snail.cs
@@ -37,7 +37,7 @@ namespace S2ObjectDefinitions.Enemies
 					{ "Right", 1 }
 				},
 				(obj) => obj.PropertyValue & 1, // in-game is a bit iffy, startup only looks at first bit but oob reset looks at entire number (== 0) - we're just gonna stick with 1st bit here
-				(obj, value) => obj.PropertyValue = (byte)((obj.PropertyValue & ~1) | (byte)((int)value)));
+				(obj, value) => obj.PropertyValue = (byte)((obj.PropertyValue & ~1) | (int)value));
 		}
 
 		public override ReadOnlyCollection<byte> Subtypes
